added convertColorSpacepipline class to test openCV changing colorspace, but need to test more (EasyOpenCVTest) using vuforia with easyopencv worked (changed testeasycv.java and ran on robot)
for movement auto:
Added the move distance from wall to the blocks, then it turns 90 degrees and readjusts to get the block and then moves and places. Need to work on EasyOpenCV and remeasuring distance to the wall perpendicular to the row of blocks for grabbing the blocks. Also write a method that turns fast and then does it slow when its around 10 degrees of the target degree. Also do some gyro stuff to make sure moving and strafing is accurate